Description
This file contains the code that uses different Azure Cognitive Services classes to transcribe, identify language and translate an audio file or several audio files. the code also includes metrics of evaluation, such as the Word Error Rate (WER) and the Matching Error Rate (MER) to evaluate the transcription, and the Bleu Score to evaluate the translation.